[lsp] Basic "language server protocol" server.
This is a first version on a language protocol server for LambdaPI
developed in the context of the internship of Ismail Lacheb by him,
Frédéric Blanqui, Emilio J. Gallego Arias, and Rodolphe Lepigre.

The server has been tested with Atom and Emacs, and supports the basic
operations. In previous versions it also supported an extension to
send proof goals as a diagnostic, but this is now currently disabled
due to upstream API issues.

Some notes:

- Lines in LSP start at 0 vs Dedukti where they start at 1.
- In addition to the base basic LSP protocol, we support
  the `documentSymbol` request.
- The server has an option to limit output to a standard LSP subset,
  see --help. This is needed by Emacs' eglot, which otherwise chokes.