[lsp] Add preliminary VSCode mode.
Many thanks to Vicent Laporte who extracted the relevant parts from
VSCoq.

The mode works, however we are back at the game of `Pure` raising
`assert false`.

I am very much afraid that the current model is unworkable; the LSP
API and the internal programming model are essentially different so it
is futile to manually ensure that they are in sync. Indeed, almost
every time I get back to the code, I tend to find some new assert that
is triggered by the LSP code.

I have added a workaround for the lack of location data on a "file not
found by require" error, but this should be fixed upstream.