Fixed hovers and go to definition (#434)
* Hovers and go to definition

* New implem for token maps, lplib and fixed pr

* Removed useless "Lp_doc" in lp_lsp

* Removed unicity check from token maps for performance

* invalid_arg in interval of popt

* New implem for token maps

* Compressed comparison implem

* Small corrections (highlight range for hover missed one character)

* Cleaned code

* Added doc, cleaned code.

* Fixed some more comment syntax.

* Added brackets in doc.

* [lplib] Apply OCamlformat

Style was a bit inconsistent with the rest of lp-lsp code, so I choose
to apply ocamlformat by hand, comments are a bit better IMHO.

No other changes.

* [lplib] [range] Refactor and split modules

We refactor the range map implementation to follow OCaml library
patterns, in particular w.r.t. interfaces. We now have two different
modules, `Range` and `RangeMap`, plus the associated functor.

* [lib] Cleanup of common utility functions.

We move the utility library to the `Lplib` library introduced in the
work on LSP. We use a system similar to the `Core` family of libs, so
our own Stdlib overlay is used by `open! Lplib`

This could be tweaked in many way, feel free to make more progress.

* [pure] Split command analysis to its own module.

That'd be a candidate for inclusion.

* [lsp_doc] Minor cleanup

* [lsp_lp] minor cleanup

We remove some utility fucntions, nit on coding style, etc...

* sanity

Co-authored-by: Emilio Jesus Gallego Arias <e+git@x80.org>
Co-authored-by: Frédéric Blanqui <frederic.blanqui@inria.fr>