[lsp] Don't error fatally for symbols without a location.
This was some leftover cleanup from #434 , the code here was not
robust.

We should anyways overhaul the whole error handling code of the server
when we have a bit of time, as it is pretty primitive owning to its
prototypical nature. Errors like this should not be so silent.

Fixes #451 , fixes #452