New typechecker, refinement, placeholders, coercions (#696)
* New typechecking algorithm

* Refinement-compatible typechecker

Typechecker can modify the type-checked term

* Add placeholders

Placeholders stand for yet unknown terms. Unlike meta-variables, they
are not typed and are not equipped with an explicit substitution. They
can't appear non linearly. Placheholders are transformed into
metavariables during typechecking. Metavariables may still be used,
but they must have been generated previously by the type checker. They
are identified by a key only.

* Optimise type checker

avoid building type checked term when possible: if term is not modified,
the input term is returned, rather than building it.

* Second optim: lifted contexts

keep a copy of contexts where types are already lifted

* Removing unused variables warnings

Whether a variable is used or not in a term cannot be decided while
scoping because placeholders does not depend on variables (in the
Bindlib sense) but may be transformed (during type checking) into
metavariables that depend on all them.

* Removing metavariable name

Metavariable are only identified by their key

* Changelog