Change the way lsp server is launched from Vscode extension (#1164)
As described in issue #1163 , one may want to launch the `Lambdapi Lsp server` from the `Vscode` extension  using a custom command. This is especially useful when installing the `Lambdapi extension` on `Windows` where `Lambdapi` is launched through `wsl`. This is done in this PR by allowing to read the arguments of the launch command as well as the command it self, from the `package.json` file instead of wiring it in the code.