Emacs/del outdated files (#1188)
This PR subsumes #1170 by @tarsius.

The information in <name>-pkg.el did not agree with the information in <name>.el. This pull-requests addresses that be removing the outdated <name>-pkg.el.

However, since <name>-pkg.el is necessary to install the package locally for debugging, an entry is added to the make file to generate it and include it to the `tar` file for local testing only.