Vscode/fix follow cursor (#1213)
This PR fixes the problem of the cursor mode where proofs are navigated following the cursor position. The problem is that this mode checks proofs even when non Lambdapi files are navigated resulting in highlighting the text (mostly in red) in every file that is open in Vscode.

Besides, this PR adds a rule in the vscode/Makefile to install the lambdapi Vscode extension from local code essentially for debugging and beta testing the extension and it updates the package-lock.json file in editors/vscode.

Finally, the extension is released to version 0.2.4.