Run the Vscode job with different Nodejs, vsce and @types/vscode version (#1256)
This PR, modifies the github pipiline to build the VScode extension with different versions of Nodejs, vsce and @types/vscode. It also uses the latest version of Nodejs for which vscode has been fixed

Moreover this PR separates the pipeline workflow in two files to trigger only the workflow for impacted part of the Repo (whether Lambdapi or Vscode extension).

We still have to decide what version of types_vscode, vsce and Nodejs we want to test with the pipeline.