Websearch switched to rocq{Lexer,Parser}
- doc and welcome page of websearch to be updated with the new syntax
- in "lambdapi search" and VScode -> and forall cannot be used any more
- the new syntax allows all the LP syntax plus:
  t ::= ... | forall binders, t | exists binders, t | fun binders => t
            | t -> t
- the binders used in the new entries allow also "x ... z : t" that is
  currently rejected by LP (but it is allowed in Coq)

TODO: but for exists, all the Coq's stdlib notations are NOT in place
(e.g. arithmetic/logic connectives, numbers, ...). How should we add them?
Maybe one possible path is to have "special" notation files that can be fed
to websearch when it starts. As for rewriting rules for normalization, the
notation files are supposed to add notation to symbols that have NOT been
put in the environment. It may be feasible: the file that calls Pratter
already uses our find_sym to resolve symbols!