Emacs test pipeline (#1301)
This PR creates a workflow to test the Emacs package with different version of Eglot, Math-symbol-lists and Highlight.
Note the script used for this purpose hacks the installation process of Elpa to install specific versions of the dependencies because Elpa does not allow as far as I know to install specific (previous) versions of packages.