Remove duplicated position in diagnostics severity 1 (#1322)
This PR fixes issue https://github.com/Deducteam/lambdapi/issues/1311.
To do so, the position of the error in diagnostics is not included in the message when such error is in the file currently open in the editor.
When the error occurs in a different file ("imported" with `open` command for instance) the position is included in the diagnostic message.
The position is also maintained in the messages printed in the console.
Finally, hover messages are now formatted in Markdown to allow syntax highlighting.